backend/requirements.txt
# This file contains the project dependencies.
# Each line represents a package to be installed.
# The following packages are needed: flask, itsdangerous, Werkzeug, SQLAlchemy, alembic and psycopg2-binary

alembic==1.15.1
Flask==3.1.0
Flask-Limiter==3.5.0
flask_cors==5.0.1
flask_jwt_extended==4.7.1
flask_migrate==4.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
Werkzeug==3.1.3
argon2-cffi==23.1.0  # Für sicheres Passwort-Hashing
gunicorn   # Hinzugefügt
PyMySQL    # Hinzugefügt
